requirements.txt
# Core dependencies for VQ-VAE research

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Experiment tracking
wandb>=0.15.0

# Data and visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scipy>=1.10.0

# Evaluation metrics
clean-fid>=0.1.35           # For r-FID computation
Pillow>=9.0.0               # Image processing for FID

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Development (optional)
pytest>=7.3.0
black>=23.0.0
